# [libc++] Optimize the std::mismatch tail