@vitalybuka 2024-03-31
> @vitalybuka Where can I find the test that is failing?

Isn't IncludeInserterTest.InsertCategoryHeaderObjectiveC the test? 
https://github.com/llvm/llvm-project/blob/82c6eeed08b1c8267f6e92d594c910fe57a9775e/clang-tools-extra/unittests/clang-tidy/IncludeInserterTest.cpp#L763 ?

If you have Ubuntu/Debian host, you can fully reproduce the bot with  https://github.com/google/sanitizers/wiki/SanitizerBotReproduceBuild