@vitalybuka 2024-04-01
Then it's easies to check with assert or just try to fix without reproducing with asan build?