@vitalybuka 2024-04-01
Confirmed with `assert(RHS.size() < LHS.size())` https://github.com/llvm/llvm-project/pull/87213